correct errors in files

+++ 0x08-recursion/5-sqrt_recursion.c
@@ -19,7 +19,7 @@ int binary_search(int n, int low, int high)
 	if (low <= high)
 	{
 		int mid = (low + high) / 2;
-		long long square =(long long) mid * mid;
+		int square = mid * mid;
 
 		if (square == n)
 		{
